Ended the general structure and made it work
Ended to implement the functions and classes that allow to make it easier for te core engine. Now, we can have many planets of the solar system and display it all easily. Changed the location of the elements in the 3D scenes just to seee it well separated. Finally, it is also added better textures versions